# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoalConnect Backend API
  version: 1.0.0
  description: |
    ## GoalConnect REST API

    Digital Talent Scouting Platform for Ethiopian Football.

    ### Roles
    | Role    | Description |
    |---------|-------------|
    | admin   | Platform administrator |
    | academy | Football academy that manages players |
    | scout   | Talent scout who discovers players |
    | player  | Individual footballer |

    ### Authentication
    All protected routes require a **Bearer JWT** token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    Tokens are obtained from `/auth/login` or `/auth/register`.

    ### Real-time Messaging (Socket.IO)
    Core messaging is handled via **Socket.IO**. For local development use `ws://localhost:5000`; in production use `wss://goalconnect-backend-repo-2.onrender.com`.

    Connect with JWT in the handshake (local example):
    ```js
    const socket = io('http://localhost:5000', {
      auth: { token: '<jwt>' }
    });
    ```

    Production / deployed example:
    ```js
    const socket = io('https://goalconnect-backend-repo-2.onrender.com', {
      transports: ['websocket'],
      auth: { token: '<jwt>' }
    });
    ```
    **Client → Server events:**
    - `message:send` — `{ toUserId, content, sharedVideo? }` (send a message; supports ack callback)
    - `conversation:history` — `{ withUserId, limit? }` (fetch recent messages; ack returns data)
...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-f\d]{24}$'
      example: 680df8bf8f0e823a9ca20e99
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                example: field
              msg:
                type: string
                example: Email is required
              path:
                type: string
                example: email
              location:
                type: string
                example: body
...
    Video:
      type: object
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        player:
          oneOf:
            - $ref: "#/components/schemas/ObjectId"
            - type: object
              additionalProperties: true
        match:
          $ref: "#/components/schemas/ObjectId"
        uploadedBy:
          $ref: "#/components/schemas/ObjectId"
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
        videoType:
          type: string
          enum: [highlight, drill, match]
        drillType:
          type: string
          enum: [dribbling, shooting, passing, speed, agility, other]
        privacy:
          type: string
          enum: [public, scout_only, private]
        processingStatus:
          type: string
          enum: [uploaded, processing, analyzed, failed]
        views:
          type: integer
...
    DrillAnalysis:
      type: object
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        video:
          $ref: "#/components/schemas/ObjectId"
        player:
          $ref: "#/components/schemas/ObjectId"
        metrics:
          type: object
          additionalProperties: true
        heatmapData:
          type: object
          additionalProperties: true
        trackingData:
          type: object
          additionalProperties: true
        confidence:
          type: number
...
paths:
...
  /players/{id}/videos:
    get:
      tags: [Players]
      summary: Get videos for a player (privacy-filtered)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1 }
        - in: query
          name: videoType
          schema: { type: string, enum: [highlight, drill, match] }
      responses:
        "200":
          description: Player videos
...
  /videos/feed:
    get:
      tags: [Videos]
      summary: Smart discovery feed (role-filtered)
      description: |
        Returns a paginated list of `FeedItem` objects.
        Visibility rules by role:
        - **Admin** — all analyzed or uploaded videos
        - **Academy** — all public player videos + their own uploaded videos
        - **Scout / Player / Guest** — public videos that are linked to a player

        Pass the JWT in the `Authorization` header to get role-based filtering.
        Omit it to receive only public, player-linked videos.
      security: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
          description: Page number
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
          description: Items per page
      responses:
        "200":
          description: Feed items list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 10
                  total:
                    type: integer
                    example: 47
                  page:
                    type: integer
                    example: 1
                  pages:
                    type: integer
                    example: 5
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
  /videos:
    get:
      tags: [Videos]
      summary: List public analyzed videos
      description: |
        Returns only videos where `privacy: public` AND `processingStatus: analyzed`.
        Use `/videos/feed` for a broader result set that includes uploaded-but-not-yet-analyzed videos.
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, default: 20 }
        - in: query
          name: videoType
          schema: { type: string, enum: [highlight, drill, match] }
          description: Filter by video type
        - in: query
          name: playerId
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Filter to a specific player's videos
      responses:
        "200":
          description: Video list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  count: { type: integer }
                  total: { type: integer }
                  page: { type: integer }
                  pages: { type: integer }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Video"
    post:
      tags: [Videos]
      summary: Upload a video (academy or player)
      description: |
        Uploads a video file to **Cloudinary** and creates a `Video` document.

        **Content-Type must be `multipart/form-data`.**

        After upload the video is in `processingStatus: uploaded`.
        A future CV pipeline will update it to `analyzed`.

        **Who can upload:**
        - **Academy** — can upload for any player in their roster (pass `playerId`), or as a generic academy video (omit `playerId`)
        - **Player** — always uploads to their own profile; `playerId` is ignored

        **Requires:** `approved` account status.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadVideoRequest"
      responses:
        "201":
          description: Video uploaded and queued for processing
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    type: string
                    example: Video uploaded successfully. It will be processed for analysis.
                  data:
                    $ref: "#/components/schemas/Video"
        "400":
          description: Validation failed or no file attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                noFile:
                  summary: No video file in request
                  value:
                    success: false
                    message: Video file is required
                validationFail:
                  summary: Missing required field
                  value:
                    success: false
                    message: Validation failed
                    errors:
                      - type: field
                        msg: Video type is required
                        path: videoType
                        location: body
        "401":
          description: Missing or invalid token
        "403":
          description: Account not approved or wrong role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                success: false
                message: Your account is pending approval. Please wait for admin verification.
  /videos/{id}:
    get:
      tags: [Videos]
      summary: Get video by ID
      description: |
        Fetches a single video and increments the view counter.

        **Privacy rules:**
        - `public` — anyone can view
        - `scout_only` — requires authenticated scout or admin
        - `private` — requires the owning academy or admin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Video document ID
      responses:
        "200":
          description: Video details (view count incremented)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: "#/components/schemas/Video"
        "403":
          description: Video is private or scout-only and caller lacks permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                privateVideo:
                  value: { success: false, message: This video is private }
                scoutOnly:
                  value: { success: false, message: This video is only available to scouts }
        "404":
          description: Video not found
    put:
      tags: [Videos]
      summary: Update video metadata
      description: |
        Updates `title`, `description`, `privacy`, or `drillType`.
        Cannot change `videoUrl` or `videoType` after upload.

        **Who can update:**
        - The **uploader** (whoever called POST /videos)
        - An **academy** user if the video belongs to one of their players
        - Any **admin**
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVideoRequest"
            example:
              title: Updated — Abebe Sprint Drill
              privacy: public
      responses:
        "200":
          description: Video updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Video updated successfully }
                  data:
                    $ref: "#/components/schemas/Video"
        "403":
          description: Not the owner or authorized role
        "404":
          description: Video not found
    delete:
      tags: [Videos]
      summary: Delete video (owner only)
      description: |
        Deletes the video from the database **and** removes the file from Cloudinary.
        Also deletes any associated `DrillAnalysis` document.

        **Who can delete:**
        - The **uploader**
        - An **academy** user if the video belongs to one of their players
        - Any **admin**
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: Video deleted from DB and Cloudinary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Video deleted successfully }
        "403":
          description: Not the owner or authorized role
        "404":
          description: Video not found
  /videos/{id}/analysis:
    get:
      tags: [Videos]
      summary: Get CV drill analysis for a video
      description: |
        Returns the `DrillAnalysis` document for a video.
        Only available after the CV pipeline has run (`processingStatus: analyzed`).
        Includes speed metrics, distance, agility/stamina/technical scores, and heatmap data.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: Analysis data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: "#/components/schemas/DrillAnalysis"
        "404":
          description: Video or analysis not found
  /videos/{id}/view:
    post:
      tags: [Videos]
      summary: Increment video view count
      description: Call this when the player's video starts playing in the client. No auth required.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: View count updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      views: { type: integer, example: 143 }
        "404":
          description: Video not found
  /videos/{id}/like:
    post:
      tags: [Videos]
      summary: Toggle like on a video
      description: |
        Toggles the current user's like state for a video.

        - If the user has not liked the video yet, their User `_id` is appended to `Video.likes`.
        - If they have already liked it, their User `_id` is removed from `Video.likes`.

        This endpoint is safe to click repeatedly while testing the UI because it flips between
        `Video liked` and `Video unliked`.

        **Auth:** any authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    type: string
                    example: Video liked
                    enum: [Video liked, Video unliked]
                  data:
                    type: object
                    properties:
                      likes:
                        type: array
                        items:
                          $ref: "#/components/schemas/ObjectId"
                        description: Updated array of User IDs who liked the video
                      likesCount:
                        type: integer
                        example: 14
              examples:
                liked:
                  summary: User liked the video
                  value:
                    success: true
                    message: Video liked
                    data:
                      likes:
                        - 680df8bf8f0e823a9ca20e99
                        - 681abc123def456789012345
                      likesCount: 2
                unliked:
                  summary: User unliked the video
                  value:
                    success: true
                    message: Video unliked
                    data:
                      likes:
                        - 680df8bf8f0e823a9ca20e99
                      likesCount: 1
        "401":
          description: Unauthorized
        "404":
          description: Video not found
  /videos/{id}/comments:
    get:
      tags: [Videos]
      summary: Get top-level comments for a video (with nested replies)
      description: |
        Returns all **top-level** comments (where `parentComment` is null).
        Each comment includes its `replies` array populated with reply objects.
        Results are sorted newest-first.

        Use this endpoint first when testing the comment UI, then call the POST endpoint
        to add a top-level comment or reply.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Video ID
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  count: { type: integer, example: 3 }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
              examples:
                sample:
                  summary: Example comment list response
                  value:
                    success: true
                    count: 1
                    data:
                      - _id: 680df8bf8f0e823a9ca20e99
                        text: Great footwork!
                        user:
                          _id: 681abc123def456789012345
                          fullName: John Scout
                          profileImageUrl: https://res.cloudinary.com/...
                          role: scout
                        video: 682abc123def456789012345
                        likes:
                          - 680df8bf8f0e823a9ca20e99
                        parentComment: null
                        replies: []
                        createdAt: 2026-05-19T08:30:00.000Z
    post:
      tags: [Videos]
      summary: Add a comment or reply to a video
      description: |
        Posts a new comment on the video.

        To post a **reply**, include `parentComment` with the `_id` of the comment being replied to.
        The reply is automatically added to the parent's `replies` array.

        **Auth:** any authenticated user.

        If `parentComment` is omitted, a top-level comment is created.
        If `parentComment` is present, the comment is created as a reply and linked back to the parent.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Video ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
            examples:
              topLevel:
                summary: Top-level comment
                value:
                  text: Incredible technique, great future ahead!
              reply:
                summary: Reply to an existing comment
                value:
                  text: Totally agree!
                  parentComment: 680df8bf8f0e823a9ca20e99
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: "#/components/schemas/Comment"
              examples:
                topLevel:
                  summary: Created top-level comment
                  value:
                    success: true
                    data:
                      _id: 682abc123def456789012345
                      text: Incredible technique, great future ahead!
                      user:
                        _id: 681abc123def456789012345
                        fullName: John Scout
                        profileImageUrl: https://res.cloudinary.com/...
                        role: scout
                      video: 680df8bf8f0e823a9ca20e99
                      likes: []
                      parentComment: null
                      replies: []
                      createdAt: 2026-05-19T08:30:00.000Z
                reply:
                  summary: Created reply
                  value:
                    success: true
                    data:
                      _id: 682abc123def456789012346
                      text: Totally agree!
                      user:
                        _id: 681abc123def456789012345
                        fullName: John Scout
                        profileImageUrl: https://res.cloudinary.com/...
                        role: scout
                      video: 680df8bf8f0e823a9ca20e99
                      likes: []
                      parentComment: 680df8bf8f0e823a9ca20e99
                      replies: []
                      createdAt: 2026-05-19T08:31:00.000Z
        "400":
          description: Empty comment text or parent not found
        "401":
          description: Unauthorized
        "404":
          description: Video not found
  /videos/{id}/comments/{commentId}:
    delete:
      tags: [Videos]
      summary: Delete a comment
      description: |
        Only the **comment author** or an **admin** can delete a comment.
        Note: Deleting a parent comment does **not** cascade-delete its replies automatically.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Video ID
        - in: path
          name: commentId
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Comment ID to delete
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: Comment deleted successfully }
        "403":
          description: Not the comment author or admin
        "404":
          description: Comment not found
  /videos/{id}/comments/{commentId}/like:
    post:
      tags: [Videos]
      summary: Toggle like on a comment
      description: |
        Adds or removes the current user's like from a comment.

        - If the user has not liked the comment yet, their User `_id` is added to `Comment.likes`.
        - If they already liked it, their User `_id` is removed.

        The response includes the updated `likes` array and `likesCount` for immediate UI updates.

        **Auth:** any authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Video ID
        - in: path
          name: commentId
          required: true
          schema:
            $ref: "#/components/schemas/ObjectId"
          description: Comment ID to like/unlike
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    type: string
                    enum: [Comment liked, Comment unliked]
                    example: Comment liked
                  data:
                    type: object
                    properties:
                      likes:
                        type: array
                        items:
                          $ref: "#/components/schemas/ObjectId"
                      likesCount:
                        type: integer
                        example: 5
              examples:
                liked:
                  summary: Comment liked
                  value:
                    success: true
                    message: Comment liked
                    data:
                      likes:
                        - 680df8bf8f0e823a9ca20e99
                        - 681abc123def456789012345
                      likesCount: 2
                unliked:
                  summary: Comment unliked
                  value:
                    success: true
                    message: Comment unliked
                    data:
                      likes:
                        - 680df8bf8f0e823a9ca20e99
                      likesCount: 1
        "401":
          description: Unauthorized
        "404":
          description: Comment not found
...