# backend/src/routes/video.routes.js
const express = require("express");
const { body } = require("express-validator");
const router = express.Router();

const {
  getVideos,
  getVideo,
  uploadVideo,
  uploadVideoFileOnly,
  updateVideo,
  deleteVideo,
  getVideoAnalysis,
  upsertTrackerAnalytics,
  getVideoFeed,
  incrementVideoView,
  toggleVideoLike,
  getVideoComments,
  addVideoComment,
  deleteVideoComment,
  toggleCommentLike,
  getAnalysisStatus,
  updateAnalysisStatus,
  getTrackerAnalyticsReview,
  assignTrackerAnalytics,
} = require("../controllers/video.controller");

const { protect, optionalAuth } = require("../middleware/auth.middleware");
const { authorize, requireApproved } = require("../middleware/role.middleware");
const {
  uploadVideoMw,
  withUploadErrorHandling,
} = require("../middleware/upload.middleware");

// Validation rules
const uploadVideoValidation = [
  body("playerId").optional().isMongoId().withMessage("Invalid player ID"),
  body("title")
    .notEmpty()
    .withMessage("Title is required")
    .isLength({ min: 2, max: 100 })
    .withMessage("Title must be between 2 and 100 characters"),
  body("description")
    .optional()
    .isLength({ max: 500 })
    .withMessage("Description cannot exceed 500 characters"),
  // videoUrl comes from Cloudinary via multer, so we remove the requirement here.
  body("videoType")
    .notEmpty()
    .withMessage("Video type is required")
    .isIn(["highlight", "drill", "match"])
    .withMessage("Video type must be highlight, drill, or match"),
  body("drillType")
    .optional()
    .isIn(["dribbling", "shooting", "passing", "speed", "agility", "other"])
    .withMessage("Invalid drill type"),
  body("privacy")
    .optional()
    .isIn(["public", "scout_only", "private"])
    .withMessage("Privacy must be public, scout_only, or private"),
];

const updateVideoValidation = [
  body("title")
    .optional()
    .isLength({ min: 2, max: 100 })
    .withMessage("Title must be between 2 and 100 characters"),
  body("description")
    .optional()
    .isLength({ max: 500 })
    .withMessage("Description cannot exceed 500 characters"),
  body("privacy")
    .optional()
    .isIn(["public", "scout_only", "private"])
    .withMessage("Privacy must be public, scout_only, or private"),
];

/**
 * @swagger
 * tags:
 *   name: Videos
 *   description: Player videos and analysis
 */

/**
 * @swagger
 * /videos/feed:
 *   get:
 *     summary: Get personalized video feed
 *     tags: [Videos]
 *     responses:
 *       200:
 *         description: Video feed
 */
// Public/Smart routes
router.get("/feed", optionalAuth, getVideoFeed);

/**
 * @swagger
 * /videos:
 *   get:
 *     summary: List public analyzed videos
 *     tags: [Videos]
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *       - in: query
 *         name: videoType
 *         schema:
 *           type: string
 *       - in: query
 *         name: playerId
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: List of videos
 */
router.get("/", optionalAuth, getVideos);

/**
 * @swagger
 * /videos/{id}:
 *   get:
 *     summary: Get video by ID (with privacy checks)
 *     tags: [Videos]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Video details
 *       403:
 *         description: Forbidden due to privacy
 *       404:
 *         description: Video not found
 */
router.get("/:id", optionalAuth, getVideo);

/**
 * @swagger
 * /videos/{id}/analysis:
 *   get:
 *     summary: Get analysis data for a video
 *     tags: [Videos]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Video analysis
 *       404:
 *         description: Video or analysis not found
 */
router.get("/:id/analysis", optionalAuth, getVideoAnalysis);

router.post(
  "/:id/tracker-analytics",
  protect,
  authorize("academy", "admin"),
  requireApproved,
  upsertTrackerAnalytics,
);

// router.get(
//   "/:id/tracker-analytics/review",
//   protect,
//   authorize("academy", "admin"),
//   requireApproved,
//   getTrackerAnalyticsReview, // Function not defined in controller
// );

// router.post(
//   "/:id/tracker-analytics/assign",
//   protect,
//   authorize("academy", "admin"),
//   requireApproved,
//   assignTrackerAnalytics, // Function not defined in controller
// );

// // Auto-trigger route for CV job completion (webhook/internal)
// router.post(
//   "/:id/auto-tracker-analytics",
//   // This route can be called by internal services/webhooks
//   // Add proper authentication for your CV processing service
//   autoCreateTrackerAnalytics, // Function not defined in controller
// );

/**
 * @swagger
 * /videos/{id}/view:
 *   post:
 *     summary: Increment video view count
 *     tags: [Videos]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: View count incremented
 *       404:
 *         description: Video not found
 */
router.post("/:id/view", incrementVideoView);

router.post("/:id/like", protect, requireApproved, toggleVideoLike);

router.get("/:id/comments", optionalAuth, getVideoComments);

router.post(
  "/:id/comments",
  protect,
  requireApproved,
  addVideoComment,
);

router.post(
  "/:id/comments/:commentId/like",
  protect,
  requireApproved,
  toggleCommentLike,
);

router.delete(
  "/:id/comments/:commentId",
  protect,
  requireApproved,
  deleteVideoComment,
);

/**
 * @swagger
 * /videos:
 *   post:
 *     summary: Upload a video (academy or player)
 *     tags: [Videos]
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         multipart/form-data:
 *           schema:
 *             type: object
 *             properties:
 *               video:
 *                 type: string
 *                 format: binary
 *     responses:
 *       201:
 *         description: Video uploaded
 *       400:
 *         description: Validation error
 *       401:
 *         description: Unauthorized
 *       403:
 *         description: Forbidden (requires approved account)
 */
// Protected routes (Academy and Player)
router.post(
  "/upload-file",
  protect,
  authorize("academy", "player"),
  requireApproved,
  withUploadErrorHandling(uploadVideoMw.single("video")),
  uploadVideoFileOnly,
);

router.post(
  "/",
  protect,
  authorize("academy", "player"),
  requireApproved,
  withUploadErrorHandling(uploadVideoMw.single("video")),
  uploadVideoValidation,
  uploadVideo,
);

/**
 * @swagger
 * /videos/{id}:
 *   put:
 *     summary: Update video metadata (academy only)
 *     tags: [Videos]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *     responses:
 *       200:
 *         description: Video updated
 *       401:
 *         description: Unauthorized
 *       404:
 *         description: Video not found
 */
router.put(
  "/:id",
  protect,
  authorize("academy", "player", "admin"),
  updateVideoValidation,
  updateVideo,
);

/**
 * @swagger
 * /videos/{id}:
 *   delete:
 *     summary: Delete video (owner only)
 *     tags: [Videos]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Video deleted
 *       401:
 *         description: Unauthorized
 *       404:
 *         description: Video not found
 */
router.delete(
  "/:id",
  protect,
  authorize("academy", "player", "admin"),
  deleteVideo,
);

router.get(
  "/:id/tracker-analytics/review",
  protect,
  authorize("academy", "admin"),
  requireApproved,
  getTrackerAnalyticsReview,
);

router.post(
  "/:id/tracker-analytics/assign",
  protect,
  authorize("academy", "admin"),
  requireApproved,
  assignTrackerAnalytics,
);

router.get(
  "/:id/analysis-status",
  protect,
  authorize("academy", "admin"),
  getAnalysisStatus,
);

// CV service pushes stage updates via this callback (authenticated with the token it received)
router.put("/:id/analysis-status", protect, updateAnalysisStatus);

module.exports = router;